requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.53.1
importlib_resources==6.4.5
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==1.26.4
packaging==24.1
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
scipy==1.13.1
six==1.16.0
torch==1.9.1
torchaudio==0.9.1
torchvision==0.10.1
typing_extensions==4.12.2
zipp==3.20.2
